+++ mahos/inst/overlay/iodmr_sweeper.py
@@ -69,7 +69,7 @@ class IODMRSweeperCommand(InstrumentOverlay):
 
     # Standard API
 
-    def configure(self, params: dict, label: str = "", group: str = "") -> bool:
+    def configure(self, params: dict, label: str = "") -> bool:
         if not self.check_required_params(
             params, ("start", "stop", "num", "power", "exposure_delay", "exposure_time")
         ):