+++ mahos/inst/sg.py
@@ -431,7 +431,7 @@ class N5182B(VisaInstrument):
             self.logger.error("Unknown set() key.")
             return False
 
-    def configure(self, params: dict, label: str = "", group: str = "") -> bool:
+    def configure(self, params: dict, label: str = "") -> bool:
         if not self.check_required_params(params, ("mode",)):
             return False
 
@@ -886,7 +886,7 @@ class MG3710E(VisaInstrument):
             self.logger.error("Unknown set() key.")
             return False
 
-    def configure(self, params: dict, label: str = "", group: str = "") -> bool:
+    def configure(self, params: dict, label: str = "") -> bool:
         if not self.check_required_params(params, ("mode",)):
             return False
 
@@ -1131,7 +1131,7 @@ class DS_SG(VisaInstrument):
             self.logger.error("Unknown set() key.")
             return False
 
-    def configure(self, params: dict, label: str = "", group: str = "") -> bool:
+    def configure(self, params: dict, label: str = "") -> bool:
         if not self.check_required_params(params, ("mode",)):
             return False